Create auth model
